server/pyproject.toml
[project]
name = "image-edit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.2.1",
    "diffusers>=0.32.1",
    "fastapi>=0.115.6",
    "pillow>=11.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "uvicorn>=0.34.0",
]
